// frontends/main/src/app/core/model/field.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Workflow } from './workflow';
import { PossibleValue } from './possible-value';
import { FieldModel } from './field-model';
import { WorkflowStatus } from './workflow-status';


/**
 * Fields associated with the dataset
 */
export interface Field { 
    /**
     * Scope reference
     */
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    /**
     * Event reference
     */
    eventPk?: number;
    eventId?: string;
    eventShortname?: string;
    /**
     * Dataset reference
     */
    datasetPk: number;
    datasetId: string;
    datasetModelId: string;
    /**
     * Field pk
     */
    pk: number;
    creationTime: Date;
    lastUpdateTime: Date;
    model: FieldModel;
    /**
     * Field model id
     */
    modelId: string;
    /**
     * Is the field attached to a removed entity?
     */
    inRemoved: boolean;
    /**
     * Is the field attached to a locked entity?
     */
    inLocked: boolean;
    /**
     * Possible values of the field, if the field type is a multiple
     */
    possibleValues: Array<PossibleValue>;
    value?: string;
    valueLabel?: string;
    newContent?: boolean;
    filePk?: number;
    fileName?: string;
    workflowStatuses: Array<WorkflowStatus>;
    possibleWorkflows: Array<Workflow>;
}