// frontends/main/src/app/core/model/form.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FormModel } from './form-model';
import { Workflow } from './workflow';
import { WorkflowStatus } from './workflow-status';


/**
 * A form is a visual grouping of fields.
 */
export interface Form { 
    /**
     * Scope reference
     */
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    /**
     * Event reference
     */
    eventPk?: number;
    eventId?: string;
    eventShortname?: string;
    /**
     * Form pk
     */
    pk: number;
    creationTime: Date;
    lastUpdateTime: Date;
    model: FormModel;
    /**
     * ID of the form model
     */
    modelId: string;
    removed: boolean;
    /**
     * Is the form attached to a removed entity?
     */
    inRemoved: boolean;
    /**
     * Is the form attached to a locked entity?
     */
    inLocked: boolean;
    /**
     * Can the user update the form?
     */
    canWrite: boolean;
    /**
     * Can the user remove the form?
     */
    canBeRemoved: boolean;
    printable: boolean;
    printButtonLabel?: { [key: string]: string; };
    workflowStatuses: Array<WorkflowStatus>;
    possibleWorkflows: Array<Workflow>;
}