// frontends/main/src/app/core/model/event.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventModel } from './event-model';
import { Workflow } from './workflow';
import { WorkflowStatus } from './workflow-status';


export interface Event { 
    /**
     * Scope reference
     */
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    /**
     * Event pk
     */
    pk: number;
    id: string;
    eventGroupNumber: number;
    creationTime: Date;
    lastUpdateTime: Date;
    model: EventModel;
    /**
     * Event model id
     */
    modelId: string;
    shortname: string;
    longname: string;
    removed: boolean;
    /**
     * Is the event attached to a removed entity?
     */
    inRemoved: boolean;
    locked: boolean;
    /**
     * Is the event attached to a locked entity?
     */
    inLocked: boolean;
    /**
     * Can the user update the event?
     */
    canWrite: boolean;
    /**
     * Can the user remove the event?
     */
    canBeRemoved: boolean;
    expectedDate: Date;
    date: Date;
    endDate?: Date;
    mandatory: boolean;
    expected: boolean;
    notDone: boolean;
    blocking: boolean;
    workflowStatuses: Array<WorkflowStatus>;
    possibleWorkflows: Array<Workflow>;
}