// frontends/main/src/app/core/model/form-info.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Field } from './field';


export interface FormInfo { 
    field: Field;
    scopePk: number;
    eventPk: number;
    formPk: number;
}